+++ flutter_secure_storage/android/src/main/java/com/it_nomads/fluttersecurestorage/ciphers/RSACipher18Implementation.java
@@ -1,11 +1,12 @@
 package com.it_nomads.fluttersecurestorage.ciphers;
 
-import android.annotation.TargetApi;
 import android.content.Context;
 import android.content.res.Configuration;
 import android.os.Build;
 import android.security.keystore.KeyGenParameterSpec;
 import android.security.keystore.KeyProperties;
+
+import androidx.annotation.RequiresApi;
 
 import java.math.BigInteger;
 import java.security.Key;
@@ -21,59 +22,49 @@ import java.util.Locale;
 import javax.crypto.Cipher;
 import javax.security.auth.x500.X500Principal;
 
-class RSACipher18Implementation {
+class RSACipher18Implementation implements KeyCipher {
 
     private static final String KEYSTORE_PROVIDER_ANDROID = "AndroidKeyStore";
     private static final String TYPE_RSA = "RSA";
-    private final String KEY_ALIAS;
-    private final Context context;
+    protected final String keyAlias;
+    protected final Context context;
 
 
     public RSACipher18Implementation(Context context) throws Exception {
-        KEY_ALIAS = context.getPackageName() + ".FlutterSecureStoragePluginKey";
         this.context = context;
+        keyAlias = createKeyAlias();
         createRSAKeysIfNeeded(context);
     }
 
+    protected String createKeyAlias() {
+        return context.getPackageName() + ".FlutterSecureStoragePluginKey";
+    }
+
+    @Override
     public byte[] wrap(Key key) throws Exception {
         PublicKey publicKey = getPublicKey();
         Cipher cipher = getRSACipher();
-        cipher.init(Cipher.WRAP_MODE, publicKey);
+        cipher.init(Cipher.WRAP_MODE, publicKey, getAlgorithmParameterSpec());
 
         return cipher.wrap(key);
     }
 
+    @Override
     public Key unwrap(byte[] wrappedKey, String algorithm) throws Exception {
         PrivateKey privateKey = getPrivateKey();
         Cipher cipher = getRSACipher();
-        cipher.init(Cipher.UNWRAP_MODE, privateKey);
+        cipher.init(Cipher.UNWRAP_MODE, privateKey, getAlgorithmParameterSpec());
 
         return cipher.unwrap(wrappedKey, algorithm, Cipher.SECRET_KEY);
-    }
-
-    public byte[] encrypt(byte[] input) throws Exception {
-        PublicKey publicKey = getPublicKey();
-        Cipher cipher = getRSACipher();
-        cipher.init(Cipher.ENCRYPT_MODE, publicKey);
-
-        return cipher.doFinal(input);
-    }
-
-    public byte[] decrypt(byte[] input) throws Exception {
-        PrivateKey privateKey = getPrivateKey();
-        Cipher cipher = getRSACipher();
-        cipher.init(Cipher.DECRYPT_MODE, privateKey);
-
-        return cipher.doFinal(input);
     }
 
     private PrivateKey getPrivateKey() throws Exception {
         KeyStore ks = KeyStore.getInstance(KEYSTORE_PROVIDER_ANDROID);
         ks.load(null);
 
-        Key key = ks.getKey(KEY_ALIAS, null);
+        Key key = ks.getKey(keyAlias, null);
         if (key == null) {
-            throw new Exception("No key found under alias: " + KEY_ALIAS);
+            throw new Exception("No key found under alias: " + keyAlias);
         }
 
         if (!(key instanceof PrivateKey)) {
@@ -87,20 +78,20 @@ class RSACipher18Implementation {
         KeyStore ks = KeyStore.getInstance(KEYSTORE_PROVIDER_ANDROID);
         ks.load(null);
 
-        Certificate cert = ks.getCertificate(KEY_ALIAS);
+        Certificate cert = ks.getCertificate(keyAlias);
         if (cert == null) {
-            throw new Exception("No certificate found under alias: " + KEY_ALIAS);
+            throw new Exception("No certificate found under alias: " + keyAlias);
         }
 
         PublicKey key = cert.getPublicKey();
         if (key == null) {
-            throw new Exception("No key found under alias: " + KEY_ALIAS);
+            throw new Exception("No key found under alias: " + keyAlias);
         }
 
         return key;
     }
 
-    private Cipher getRSACipher() throws Exception {
+    protected Cipher getRSACipher() throws Exception {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
             return Cipher.getInstance("RSA/ECB/PKCS1Padding", "AndroidOpenSSL"); // error in android 6: InvalidKeyException: Need RSA private or public key
         } else {
@@ -108,11 +99,15 @@ class RSACipher18Implementation {
         }
     }
 
+    protected AlgorithmParameterSpec getAlgorithmParameterSpec() {
+        return null;
+    }
+
     private void createRSAKeysIfNeeded(Context context) throws Exception {
         KeyStore ks = KeyStore.getInstance(KEYSTORE_PROVIDER_ANDROID);
         ks.load(null);
 
-        Key privateKey = ks.getKey(KEY_ALIAS, null);
+        Key privateKey = ks.getKey(keyAlias, null);
         if (privateKey == null) {
             createKeys(context);
         }
@@ -124,39 +119,8 @@ class RSACipher18Implementation {
     private void setLocale(Locale locale) {
         Locale.setDefault(locale);
         Configuration config = context.getResources().getConfiguration();
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-            setSystemLocale(config, locale);
-            context.createConfigurationContext(config);
-        } else {
-            setSystemLocaleLegacy(config, locale);
-            setContextConfigurationLegacy(context, config);
-        }
-    }
-
-    @SuppressWarnings("deprecation")
-    private void setContextConfigurationLegacy(Context context, Configuration config) {
-        context.getResources().updateConfiguration(config, context.getResources().getDisplayMetrics());
-    }
-
-    @SuppressWarnings("deprecation")
-    private void setSystemLocaleLegacy(Configuration config, Locale locale) {
-        config.locale = locale;
-    }
-
-    @TargetApi(Build.VERSION_CODES.N)
-    private void setSystemLocale(Configuration config, Locale locale) {
         config.setLocale(locale);
-    }
-
-    @SuppressWarnings("deprecation")
-    private AlgorithmParameterSpec makeAlgorithmParameterSpecLegacy(Context context, Calendar start, Calendar end) {
-        return new android.security.KeyPairGeneratorSpec.Builder(context)
-                .setAlias(KEY_ALIAS)
-                .setSubject(new X500Principal("CN=" + KEY_ALIAS))
-                .setSerialNumber(BigInteger.valueOf(1))
-                .setStartDate(start.getTime())
-                .setEndDate(end.getTime())
-                .build();
+        context.createConfigurationContext(config);
     }
 
     private void createKeys(Context context) throws Exception {
@@ -174,16 +138,7 @@ class RSACipher18Implementation {
             if (Build.VERSION.SDK_INT < Build.VERSION_CODES.M) {
                 spec = makeAlgorithmParameterSpecLegacy(context, start, end);
             } else {
-                KeyGenParameterSpec.Builder builder = new KeyGenParameterSpec.Builder(KEY_ALIAS, KeyProperties.PURPOSE_DECRYPT | KeyProperties.PURPOSE_ENCRYPT)
-                        .setCertificateSubject(new X500Principal("CN=" + KEY_ALIAS))
-                        .setDigests(KeyProperties.DIGEST_SHA256)
-                        .setBlockModes(KeyProperties.BLOCK_MODE_ECB)
-                        .setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_RSA_PKCS1)
-                        .setCertificateSerialNumber(BigInteger.valueOf(1))
-                        .setCertificateNotBefore(start.getTime())
-                        .setCertificateNotAfter(end.getTime());
-
-                spec = builder.build();
+                spec = makeAlgorithmParameterSpec(context, start, end);
             }
 
             kpGenerator.initialize(spec);
@@ -191,5 +146,30 @@ class RSACipher18Implementation {
         } finally {
             setLocale(localeBeforeFakingEnglishLocale);
         }
+    }
+
+    // Flutter gives deprecation warning without suppress
+    @SuppressWarnings("deprecation")
+    private AlgorithmParameterSpec makeAlgorithmParameterSpecLegacy(Context context, Calendar start, Calendar end) {
+        return new android.security.KeyPairGeneratorSpec.Builder(context)
+                .setAlias(keyAlias)
+                .setSubject(new X500Principal("CN=" + keyAlias))
+                .setSerialNumber(BigInteger.valueOf(1))
+                .setStartDate(start.getTime())
+                .setEndDate(end.getTime())
+                .build();
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.M)
+    protected AlgorithmParameterSpec makeAlgorithmParameterSpec(Context context, Calendar start, Calendar end) {
+        final KeyGenParameterSpec.Builder builder = new KeyGenParameterSpec.Builder(keyAlias, KeyProperties.PURPOSE_DECRYPT | KeyProperties.PURPOSE_ENCRYPT)
+                .setCertificateSubject(new X500Principal("CN=" + keyAlias))
+                .setDigests(KeyProperties.DIGEST_SHA256)
+                .setBlockModes(KeyProperties.BLOCK_MODE_ECB)
+                .setEncryptionPaddings(KeyProperties.ENCRYPTION_PADDING_RSA_PKCS1)
+                .setCertificateSerialNumber(BigInteger.valueOf(1))
+                .setCertificateNotBefore(start.getTime())
+                .setCertificateNotAfter(end.getTime());
+        return builder.build();
     }
 }