+++ flutter_secure_storage/lib/options/linux_options.dart
@@ -1,5 +1,7 @@
-part of flutter_secure_storage;
+part of '../flutter_secure_storage.dart';
 
+/// Specific options for Linux platform.
+/// Currently there are no specific linux options available.
 class LinuxOptions extends Options {
   const LinuxOptions();