+++ flutter_secure_storage/lib/options/macos_options.dart
@@ -1,10 +1,31 @@
-part of flutter_secure_storage;
+part of '../flutter_secure_storage.dart';
 
-class MacOsOptions extends Options {
-  const MacOsOptions();
+/// Specific options for macOS platform.
+class MacOsOptions extends AppleOptions {
+  const MacOsOptions({
+    String? groupId,
+    String? accountName = AppleOptions.defaultAccountName,
+    KeychainAccessibility accessibility = KeychainAccessibility.unlocked,
+    bool synchronizable = false,
+  }) : super(
+          groupId: groupId,
+          accountName: accountName,
+          accessibility: accessibility,
+          synchronizable: synchronizable,
+        );
 
   static const MacOsOptions defaultOptions = MacOsOptions();
 
-  @override
-  Map<String, String> toMap() => <String, String>{};
+  MacOsOptions copyWith({
+    String? groupId,
+    String? accountName,
+    KeychainAccessibility? accessibility,
+    bool? synchronizable,
+  }) =>
+      MacOsOptions(
+        groupId: groupId ?? _groupId,
+        accountName: accountName ?? _accountName,
+        accessibility: accessibility ?? _accessibility,
+        synchronizable: synchronizable ?? _synchronizable,
+      );
 }